third_party/WebKit/LayoutTests/accessibility/contenteditable-hidden-div-expected.txt

test

This tests that a contenteditable element will not be ignored by accessibility.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS accessibilityController.focusedElement.isEqual(editableDiv) is false
PASS successfullyParsed is true

TEST COMPLETE
